+++ muna/cli/app.py
@@ -1,0 +1,89 @@
+#
+#   Muna
+#   Copyright © 2026 NatML Inc. All Rights Reserved.
+#
+
+import typer
+
+from ..logging import TracebackMarkupConsole
+from ..version import __version__
+
+from .auth import app as auth_app
+from .compile import compile_function
+from .deploy import deploy_function
+from .misc import cli_options
+from .predictions import create_prediction
+from .predictors import archive_predictor, delete_predictor, retrieve_predictor
+from .resources import app as resources_app
+from .triage import triage_predictor
+
+# Define CLI
+typer.main.console_stderr = TracebackMarkupConsole()
+app = typer.Typer(
+    name=f"Muna CLI {__version__}",
+    no_args_is_help=True,
+    pretty_exceptions_show_locals=False,
+    pretty_exceptions_short=True,
+    add_completion=False
+)
+
+# Add top level options
+app.callback()(cli_options)
+
+# Deploy
+app.command(
+    name="compile",
+    help="Compile a model for deployment.",
+    rich_help_panel="Deploy 🚀"
+)(compile_function)
+app.command(
+    name="predict",
+    help="Invoke a compiled model.",
+    context_settings={ "allow_extra_args": True, "ignore_unknown_options": True },
+    rich_help_panel="Deploy 🚀"
+)(create_prediction)
+app.command(
+    name="deploy",
+    help="Deploy a compiled model onto a compute cloud.",
+    rich_help_panel="Deploy 🚀"
+)(deploy_function)
+
+# Models
+app.command(
+    name="retrieve",
+    help="Retrieve a compiled model.",
+    rich_help_panel="Manage 📦"
+)(retrieve_predictor)
+app.command(
+    name="archive",
+    help="Archive a compiled model." ,
+    rich_help_panel="Manage 📦"
+)(archive_predictor)
+app.command(
+    name="delete",
+    help="Delete a compiled model.",
+    rich_help_panel="Manage 📦"
+)(delete_predictor)
+
+# Subcommands
+app.add_typer(
+    auth_app,
+    name="auth",
+    help="Login, logout, and check your authentication status.",
+    rich_help_panel="Auth 👤"
+)
+
+# Insiders
+app.command(
+    name="triage",
+    help="Triage a compile error.",
+    rich_help_panel="Insiders",
+    hidden=True
+)(triage_predictor)
+app.add_typer(
+    resources_app,
+    name="resources",
+    help="Manage prediction resources.",
+    rich_help_panel="Insiders",
+    hidden=True
+)